YOLO/requirements.txt
ultralytics==8.3.57
tqdm
tensorboard
pyyaml